requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dash==1.9.1
dash-bootstrap-components==0.8.3
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
Deprecated==1.2.7
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gunicorn==19.9.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
numexpr==2.7.1
numpy==1.18.1
pandas==1.0.1
plotly==4.5.3
PyGithub==1.46
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
retrying==1.3.3
six==1.14.0
tables==3.6.1
urllib3==1.25.8
vine==1.3.0
waitress==1.4.3
Werkzeug==1.0.0
wrapt==1.12.1
